+++ plugin/__init__.py
@@ -1,4 +1,4 @@
-from . import configuration, formatter, registry, settings, shell
+from . import configuration, formatter, registry, settings, shell, view
 
 from .formatter import Formatter
 from .registry import FormatterRegistry
