+++ plugin/formatter.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-from sublime import expand_variables, score_selector, Region, Settings, View
-
-import os
+from sublime import expand_variables, score_selector, Region, View
 
 from .configuration import Configuration
 from .shell import shell
+from .view import extract_variables
 
 
 class Formatter:
@@ -44,16 +43,3 @@ class Formatter:
         sublime.set_timeout(
             lambda: view.set_viewport_position(position, animate=False), 0
         )
-
-
-def extract_variables(view: View) -> Dict[str, str]:
-    settings = view.settings()
-    tab_size = settings.get("tab_size") or 0
-    indent = " " * tab_size if settings.get("translate_tabs_to_spaces") else "\t"
-
-    vars = view.window().extract_variables()
-    vars["tab_size"] = str(tab_size)
-    vars["indent"] = indent
-    vars.update(os.environ)
-
-    return vars
